Support 2 DPUs from same vendor in single cluster deployment
Since the VSP name takes the DPU Platform Identity which is a fixed
string per DPU model, that means no more 1 DPU per model can be deployed
in single cluster. Fix this by adding a node name to the name of the DPU
which consequently changes the name of the VSP to include the node name
in its' name.

Signed-off-by: William Zhao <wizhao@redhat.com>